Move Pair to be static inner class again
